% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sperrorest_misc.R
\name{as.tilename}
\alias{as.tilename}
\alias{tilename}
\alias{as_tilename_numeric}
\alias{as_character_tilename}
\alias{as_numeric_tilename}
\alias{as_tilename_character}
\alias{print_tilename}
\title{Alphanumeric tile names}
\usage{
as.tilename(x, ...)

\method{as.tilename}{numeric}(x, ...)

\method{as.character}{tilename}(x, ...)

\method{as.numeric}{tilename}(x, ...)

\method{as.tilename}{character}(x, ...)

\method{print}{tilename}(x, ...)
}
\arguments{
\item{x}{object of class \code{tilename}, \code{character}, or
\code{numeric} (of length 2).}

\item{...}{additional arguments (currently ignored).}
}
\value{
object of class \code{tilename}, \code{character}, or numeric
vector of length 2
}
\description{
Functions for generating and handling alphanumeric tile names of the
form \code{'X2:Y7'} as used by \code{\link[=partition_tiles]{partition_tiles()}} and
\code{\link[=represampling.tile.bootstrap]{represampling.tile.bootstrap()}}.
}
\examples{
tnm <- as.tilename(c(2,3))
tnm # 'X2:Y3'
as.numeric(tnm) # c(2,3)
}
\seealso{
\code{\link[=partition_tiles]{partition_tiles()}}, \code{\link[=represampling]{represampling()}},
\code{\link[=represampling.tile.bootstrap]{represampling.tile.bootstrap()}}
}
