% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-sptune.R
\name{check_response_type}
\alias{check_response_type}
\title{check_response_type}
\usage{
check_response_type(object = NULL, error_measure = NULL,
  perf_measures = NULL, option = FALSE)
}
\arguments{
\item{object}{fitted object}

\item{error_measure}{optional user-defined error measure}
}
\description{
Checks response type of input and sets error measure
}
\keyword{internal}
