% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hyperparameters.R
\name{plot_hyper_maxent}
\alias{plot_hyper_maxent}
\title{plot_hyper_maxent}
\usage{
plot_hyper_maxent(object = NULL, error_measure = NULL,
  feature_classes = NULL, color_palette = NULL)
}
\arguments{
\item{object}{a fitted object of \link{maxent}}

\item{error_measure}{an optional error measure to use which differs from the
default.}

\item{feature_classes}{optional user-defined numeric vector of
'feature_classes' values to plot.}

\item{color_palette}{\code{option} argument from \link{scale_color_viridis}.}
}
\description{
Plots tuning results of maxent
}
\examples{
\dontrun{

library(sperrorest)
data(maxent_pred)
data(maxent_response)
data(basque)

out <- sptune_maxent(x = maxent_pred, p = maxent_response, data = basque,
                     absence = TRUE)

plot_hyper_maxent(out)
}

}
