% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{remove_missing_levels}
\alias{remove_missing_levels}
\title{remove_missing_levels}
\usage{
remove_missing_levels(fit, test_data)
}
\arguments{
\item{fit}{fitted model on training data}

\item{test_data}{data to make predictions for}
}
\value{
data.frame with matching factor levels to fitted model
}
\description{
Accounts for missing factor levels present only in test data
but not in train data by setting values to NA
}
\examples{
foo <- data.frame(
  response = rnorm(3),
  predictor = as.factor(c("A", "B", "C"))
)
model <- lm(response ~ predictor, foo)
foo.new <- data.frame(predictor = as.factor(c("A", "B", "C", "D")))
predict(model, newdata = remove_missing_levels(fit = model, test_data = foo.new))
}
\keyword{internal}
