% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{remove_missing_levels_in_test_data}
\alias{remove_missing_levels_in_test_data}
\title{remove_missing_levels_in_test_data}
\usage{
remove_missing_levels_in_test_data(fit, test_data)
}
\arguments{
\item{fit}{fitted model on training data}

\item{test_data}{data to make predictions for}
}
\description{
Accounts for missing factor levels present only in train data
but not in test data by setting these to NA
}
\keyword{internal}
