% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sperrorest_resampling.R
\name{represampling_kmeans_bootstrap}
\alias{represampling_kmeans_bootstrap}
\title{Spatial block bootstrap at the level of spatial k-means clusters}
\usage{
represampling_kmeans_bootstrap(data, coords = c("x", "y"), repetition = 1,
  nfold = 10, nboot = nfold, seed1 = NULL, oob = FALSE, ...)
}
\arguments{
\item{data}{\code{data.frame} containing at least the columns specified
by \code{coords}}

\item{coords}{vector of length 2 defining the variables in \code{data} that
contain the x and y coordinates of sample locations.}

\item{repetition}{numeric vector: cross-validation repetitions
to be generated. Note that this is not the number of repetitions,
but the indices of these repetitions. E.g., use \code{repetition = c(1:100)}
to obtain (the 'first') 100 repetitions, and \code{repetition = c(101:200)}
to obtain a different set of 100 repetitions.}

\item{nfold}{see \link{partition_kmeans}}

\item{nboot}{see \link{represampling_factor_bootstrap}}

\item{seed1}{\code{seed1+i} is the random seed that will be used by
\link{set.seed} in repetition \code{i} (\code{i} in \code{repetition})
to initialize the random number generator before sampling from the data set.}

\item{oob}{see \link{represampling_factor_bootstrap}}

\item{...}{additional arguments to be passed to \link{partition_kmeans}}
}
\description{
\code{represampling_kmeans_bootstrap} performs a non-overlapping spatial
block bootstrap by resampling at the level of irregularly-shaped partitions
generated by \link{partition_kmeans}.
}
