% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs_int_cv.R
\name{rf_cv_err}
\alias{rf_cv_err}
\title{rf_cv_err}
\usage{
rf_cv_err(ntree = NULL, mtry = NULL, train = NULL, test = NULL,
  response = NULL, formula = NULL, rf_fun = NULL, ...)
}
\arguments{
\item{ntree}{number of trees to grow}

\item{mtry}{Number of variables randomly selected as candidates for each node split.}

\item{train}{training data}

\item{test}{testing data}

\item{response}{response variable}

\item{formula}{model formula}
}
\description{
Calculates AUROC for different gamma and cost values
}
\examples{
parti <- partition_kmeans(ecuador, nfold = 5, order.clusters = FALSE)
train <- ecuador[parti[[1]][[1]]$train, ]
test <- ecuador[parti[[1]][[1]]$test, ]
response <- "slides"

###########################################
#### randomForest
###########################################

fo <- slides ~ dem + slope + hcurv + vcurv + log.carea + cslope
out <- rf_cv_err(mtry = 3, ntree = 1000, train = train, test = test,
formula = fo, response = "slides", rf_fun = "randomForest")

###########################################
#### rfsrc
###########################################

fo <- slides ~ dem + slope + hcurv + vcurv + log.carea + cslope
out <- rf_cv_err(mtry = 3, ntree = 1000, train = train, test = test,
formula = fo, response = "slides", rf_fun = "rfsrc")

##------------------------------------------------------------
## regression
##------------------------------------------------------------

#' data(ecuador) # Muenchow et al. (2012), see ?ecuador
fo <- dem ~ slides + slope + hcurv + vcurv + log.carea + cslope
response <- "dem"

out <- rf_cv_err(mtry = 3, ntree = 1000, train = train, test = test,
formula = fo, response = "dem", rf_fun = "rfsrc")

}
\keyword{internal}
