% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptune.R
\name{sptune_maxent}
\alias{sptune_maxent}
\title{sptune_maxent}
\usage{
sptune_maxent(x = NULL, p = NULL, data = NULL, nfold = NULL,
  partition_fun = NULL, error_measure = NULL, absence = FALSE,
  parameter1 = NULL, parameter2 = NULL, tune = FALSE, train = NULL,
  test = NULL, ...)
}
\arguments{
\item{x}{Predictors. Raster* object or SpatialGridDataFrame, containing
grids with predictor variables. These will be used to extract values from
for the point locations. x can also be a data.frame, in which case each
column should be a predictor variable and each row a presence or background
record}

\item{p}{Occurrence data. This can be a data.frame, matrix, SpatialPoints
object, or a vector. If p is a data.frame or matrix it represents a set of
point locations; and it must have two columns with the first being the
x-coordinate (longitude) and the second the y-coordinate (latitude).
Coordinates can also be specified with a SpatialPoints* object}

\item{data}{dataframe.}

\item{nfold}{number of folds for cross-validation.}

\item{partition_fun}{character. Data partitioning method
(e.g. \link{partition_kmeans})}

\item{error_measure}{which error measure to use for optimization.
Default to 'RMSE' for numeric responses, 'AUROC' for binary classification
and 'error' for multiclass classiciation.}

\item{absence}{is the data of type 'presence-absence'?}

\item{parameter1}{optional user-defined vector of hyperparameter
to tune over. See details.}

\item{parameter2}{optional user-defined vector of hyperparameter
to tune over. See details.}

\item{...}{additional options passed to \code{partition_fun}.}
}
\description{
Tuning of Maxent (regularization parameter \eqn{\beta} &
feature classes) using spatial cross-validation.
}
\details{
This function tunes a Maxent model from the \link{dismo} package
using (spatial) cross-validation.

\code{error_measure} can be specified by the user, selecting one of the returned
error measures of \link{sptune_maxent}.

\code{sptune_maxent} is parallelized and runs on all possible cores.

Tuning is performed over the following ranges:
\tabular{rr}{
beta: \tab -10, -8, -6, -4, -2, 2, 4, 6, 8, 10, 12, 14, 16, 18, 20\cr
feature classes: \tab "L", "Q", "H", "T", "LQ", "HQ", "LQP", "LQT", "QHP",
"QHT", "QHPT"
}

Currently, fitting/testing for every parameter combination is only performed
on 1 of the partitioned folds. Hence, results reflect the performance on one
fold and are not averaged across folds.

WARNING: Setting \code{absence = TRUE} will alter the predicted values from
relative probabilities (presence-only) to probabilities (presence-absence).
Make sure to set the switch according to your data type. Results will be
biased otherwise!
}
\examples{
\dontrun{
data(maxent_pred)
data(maxent_response)
data(basque)

sptune_maxent(x = maxent_pred, p = maxent_response,
              data = basque, absence = TRUE)
}
}
\references{
Guillera-Arroita, G., Lahoz-Monfort, J. J., & Elith, J. (2014).
Maxent is not a presence-absence method: a comment on Thibaudet al. Methods
in Ecology and Evolution, 5(11), 1192-1197. doi:10.1111/2041-210x.12252
}
\seealso{
\link{plot_hyper_maxent}
}
\author{
Patrick Schratz, Alexander Brenning
}
