% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_int_cv.R
\name{sptune_rf}
\alias{sptune_rf}
\title{svm_tuning}
\usage{
sptune_rf(formula = NULL, data = NULL, accelerate = 1, nfold = NULL,
  partition_fun = NULL, rf_fun = "svm", ...)
}
\arguments{
\item{formula}{formula}

\item{data}{data frame}

\item{accelerate}{option to speed up tuning using less cost and gamma values.
Use \code{accelerate = 2} or \code{accelerate = 4} for test runs, but \code{accelerate = 1}
for actual analysis.}

\item{nfold}{number of folds for cross-validation}

\item{partition_fun}{method for partitioning the data
(e.g. \link{partition.kmeans})}

\item{...}{additional options passed to \link{SVM}}

\item{svm_fun}{which R svm package to use. See details.}

\item{type}{classification type of svm classifier}

\item{kernel}{kernel type of svm classifier}
}
\description{
Tuning of SVM (cost & gamma) using spatial cross-validation
}
\details{
This function tunes a support vectort machine from the \link{e1071},
\link{gmum.r}, \link{kernlab} packages using (spatial) cross-validation.

Currently this function is hard-coded to a binary response variable and AUROC
as error measure.
}
\examples{
data(ecuador) # Muenchow et al. (2012), see ?ecuador
fo <- slides ~ dem + slope + hcurv + vcurv + log.carea + cslope

sptune_rf(fo, ecuador, accelerate = 16, nfold = 5, 
partition.fun = "partition.kmeans", rf_fun = "randomForest") 

}
\author{
Alexander Brenning, Patrick Schratz
}
