% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_int_cv.R
\name{sptune_svm}
\alias{sptune_svm}
\title{svm_tuning}
\usage{
sptune_svm(formula = NULL, data = NULL, accelerate = 1, nfold = NULL,
  partition.fun = NULL, kernel = NULL, type = NULL, svm_fun = "svm",
  ...)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{accelerate}{option to speed up tuning using less cost and gamma values.
Use \code{accelerate = 2} or \code{accelerate = 4} for test runs, but \code{accelerate = 1}
for actual analysis.}

\item{partition.fun}{method for partitioning the data (e.g. \link{partition.kmeans})}

\item{int_cv_fold}{number of folds for cross-validation}

\item{additional}{options passed to \link{SVM}}
}
\description{
Tuning of SVM (cost & gamma) using spatial cross-validation
}
\details{
This function tunes a support vectort machine from the \link{e1071}, \link{gmum.r},
\link{kernlab} packages using (spatial) cross-validation.

Currently this function is hard-coded to a binary response variable and AUROC
as error measure.
}
\examples{
data(ecuador) # Muenchow et al. (2012), see ?ecuador
fo <- slides ~ dem + slope + hcurv + vcurv + log.carea + cslope

svm_tune <- svm_tuning(fo, ecuador, accelerate = 16, nfold = 5, 
partition.fun = "partition.kmeans", kernel = "radial", type = "C-classification") 

}
\author{
Alexander Brenning, Patrick Schratz
}
