% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sperrorest.R
\name{summary.sperrorestpackageversion}
\alias{summary.sperrorestpackageversion}
\title{Summarize package version information obtained by \code{sperrorest}}
\usage{
\method{summary}{sperrorestpackageversion}(object, ...)
}
\arguments{
\item{object}{\code{\link[=sperrorestpackageversion]{sperrorestpackageversion()}} object calculated by
\code{\link[=sperrorest]{sperrorest()}}}
}
\value{
character vector of length one
}
\description{
\code{summary.sperrorestpackageversion} returns the package version of sperrorest
}
