% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs_int_cv.R
\name{svm_cv_err}
\alias{svm_cv_err}
\title{svm_cv_err}
\usage{
svm_cv_err(cost = NULL, gamma = NULL, train = NULL, test = NULL,
  lhs = NULL, formula = NULL)
}
\arguments{
\item{cost}{svm cost value}

\item{gamma}{svm gamma value}
}
\description{
Calculates AUROC for different gamma and cost values
}
\keyword{internal}
