% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_int_cv.R
\name{svm_tuning}
\alias{svm_tuning}
\title{svm_tuning}
\usage{
svm_tuning(formula, data, accelerate = 1, int.cv.nfold = NULL,
  partition.fun = NULL)
}
\arguments{
\item{formula}{}

\item{data}{}

\item{accelerate}{option to speed up tuning using less cost and gamma values.
Use \code{accelerate = 2} or \code{accelerate = 4} for test runs, but \code{accelerate = 1}
for actual analysis.}

\item{partition.fun}{method for partitioning the data (e.g. \link{partition.kmeans})}

\item{int_cv_fold}{number of folds for cross-validation}
}
\description{
Tuning of SVM (cost & gamma) using spatial cross-validation
}
\details{
This function tunes a support vectort machine from the \link{e1071} package
using (spatial) cross-validation.
}
