% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sperrorest_misc.R
\name{tile_neighbors}
\alias{tile_neighbors}
\title{Determine the names of neighbouring tiles in a rectangular pattern}
\usage{
tile_neighbors(nm, tileset, iterate = 0, diagonal = FALSE)
}
\arguments{
\item{nm}{Character string or factor: name of a tile, e.g., \code{'X4:Y6'}}

\item{tileset}{Admissible tile names; if missing and \code{nm} is a factor
variable, then \code{levels(nm)} is used as a default for \code{tileset}.}

\item{iterate}{internal - do not change default: to control behaviour in an
interative call to this function.}

\item{diagonal}{if \code{TRUE}, diagonal neighbours are also considered
neighbours.}
}
\value{
Character string.
}
\description{
This based on 'counting' up and down based on the tile name.
}
