% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{transfer_parallel_output}
\alias{transfer_parallel_output}
\title{transfer_parallel_output}
\usage{
transfer_parallel_output(my_res = NULL, res = NULL, impo = NULL,
  pooled_error = NULL)
}
\description{
transfers output of parallel calls to runreps
}
\keyword{internal}
