% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sperrorest_data.R
\docType{data}
\name{maipo}
\alias{maipo}
\title{Maipo dataset from Marco Pena}
\format{An object of class \code{data.frame} with 7713 rows and 68 columns.}
\usage{
maipo
}
\description{
Maipo dataset from Marco Pena
}
\keyword{datasets}
\keyword{internal}
