% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-sptune.R
\name{maxent_cv_err}
\alias{maxent_cv_err}
\title{maxent_cv_err}
\usage{
maxent_cv_err(beta_multiplier = NULL, feature_classes = NULL,
  train = NULL, test = NULL, x = NULL, p = NULL, response = NULL,
  absence = FALSE, ...)
}
\arguments{
\item{train}{training data}

\item{test}{testing data}

\item{response}{response variable}

\item{formula}{model formula}
}
\description{
Calculates AUROC for different beta_multiplier and
feature_classes values
}
\keyword{internal}
