% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-sptune.R
\name{maxent_cv_err}
\alias{maxent_cv_err}
\title{maxent_cv_err}
\usage{
maxent_cv_err(parameter1 = NULL, parameter2 = NULL, train = NULL,
  test = NULL, x = NULL, p = NULL, absence = NULL, ...)
}
\arguments{
\item{parameter1}{optional user-defined vector of 'parameter1'
hyperparameter to tune over. See details.}

\item{parameter2}{optional user-defined vector of 'parameter2'
hyperparameter to tune over. See details.}

\item{train}{training data}

\item{test}{testing data}

\item{formula}{model formula}

\item{response}{response variable}
}
\description{
Tunes maxent hyperparameters parameter1 and
parameter2 using cross-validation with AUROC as error measure
}
\examples{
\dontrun{
data(maxent_pred)
data(maxent_response)

parti <- partition_kmeans(basque, nfold = 5)
train <- maxent_pred[parti[[1]][[1]]$train, ]
test <- maxent_pred[parti[[1]][[1]]$test, ]

out <- maxent_cv_err(parameter1 = 1,
                     parameter2 = "LQ",
                     train = train, test = test, x = maxent_pred,
                     p = maxent_response, response = "diplo01",
                     absence = TRUE)
}
}
\keyword{internal}
