% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{runfolds}
\alias{runfolds}
\title{runfolds}
\usage{
runfolds(j = NULL, currentSample = NULL, data = NULL, formula = NULL,
  model.args = NULL, par.cl = NULL, do.try = NULL, model.fun = NULL,
  error.fold = NULL, error.rep = NULL, pred.fun = NULL,
  imp.variables = NULL, imp.permutations = NULL, err.fun = NULL,
  train.fun = NULL, err.train = NULL, importance = NULL,
  currentRes = NULL, currentImpo = NULL, pred.args = NULL,
  progress = NULL, pooled.obs.train = NULL, pooled.obs.test = NULL,
  pooled.pred.train = NULL, response = NULL, is.factor.prediction = NULL,
  pooled.pred.test = NULL, coords = NULL, test.fun = NULL)
}
\description{
runfolds
}
\examples{

j <- 1 # running the first repetition of 'currentSample', normally we are 
# calling an apply call to seq_along nFolds of repetition
# see also 'runreps()'
data <- ecuador
imp.one.rep <- readRDS("inst/test-objects/imp.one.rep.rda")
currentSample <- readRDS("inst/test-objects/resamp.rda")[[1]]
currentRes <- readRDS("inst/test-objects/currentRes.rda")

runfolds_single <- runfolds(j = 1, data = ecuador, currentSample = currentSample,
formula = slides ~ dem + slope + hcurv + vcurv + log.carea + cslope, 
model.args = list(family = "binomial"), do.try = FALSE, model.fun = glm,
error.fold = TRUE, error.rep = TRUE, imp.permutations = 2, 
imp.variables = c("dem", "slope", "hcurv", "vcurv", "log.carea", "cslope"),
err.train = TRUE, importance = TRUE, currentRes = currentRes, 
pred.args = list(type = "response"), response = "slides", par.cl = 2, 
coords = c("x", "y"), progress = 1, pooled.obs.train = c(), 
pooled.obs.test = c(), err.fun = err.default)

# create list with multiple fold results
runfolds_list <- map(seq_along(1:4), function(rep) runfolds(j = rep, data = ecuador, currentSample = currentSample,
formula = slides ~ dem + slope + hcurv + vcurv + log.carea + cslope, 
model.args = list(family = "binomial"), do.try = FALSE, model.fun = glm,
error.fold = TRUE, error.rep = TRUE, imp.permutations = 2, 
imp.variables = c("dem", "slope", "hcurv", "vcurv", "log.carea", "cslope"),
err.train = TRUE, importance = TRUE, currentRes = currentRes, 
pred.args = list(type = "response"), response = "slides", par.cl = 2, 
coords = c("x", "y"), progress = 1, pooled.obs.train = c(), 
pooled.obs.test = c(), err.fun = err.default))

# MERGE SUBLISTS: ONLY CURRENTRES AND CURRENTIMPO

}
\keyword{internal}
