% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{runfolds}
\alias{runfolds}
\title{runfolds}
\usage{
runfolds(
  j = NULL,
  current_sample = NULL,
  data = NULL,
  i = NULL,
  formula = NULL,
  model_args = NULL,
  model_fun = NULL,
  pred_fun = NULL,
  imp_variables = NULL,
  imp_permutations = NULL,
  imp_sample_from = "test",
  err_fun = NULL,
  train_fun = NULL,
  importance = NULL,
  current_res = NULL,
  current_impo = NULL,
  pred_args = NULL,
  pooled_obs_train = NULL,
  pooled_obs_test = NULL,
  pooled_pred_train = NULL,
  response = NULL,
  progress = NULL,
  pooled_pred_test = NULL,
  coords = NULL,
  test_fun = NULL,
  imp_one_rep = NULL,
  do_gc = NULL,
  test_param = NULL,
  train_param = NULL
)
}
\description{
Runs model fitting, error estimation and variable importance on
fold level
}
\keyword{internal}
