% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{runreps}
\alias{runreps}
\title{runreps}
\usage{
runreps(currentSample = NULL, data = NULL, formula = NULL,
  model.args = NULL, par.cl = NULL, do.gc = NULL, imp.one.rep = NULL,
  do.try = NULL, model.fun = NULL, error.fold = NULL, error.rep = NULL,
  pred.fun = NULL, imp.variables = NULL, imp.permutations = NULL,
  err.fun = NULL, train.fun = NULL, err.train = NULL, importance = NULL,
  currentRes = NULL, currentImpo = NULL, pred.args = NULL,
  progress = NULL, pooled.obs.train = NULL, pooled.obs.test = NULL,
  pooled.pred.train = NULL, response = NULL, is.factor.prediction = NULL,
  pooled.pred.test = NULL, coords = NULL, test.fun = NULL,
  par.mode = NULL, i = NULL)
}
\description{
Runs model fitting, error estimation and variable importance on fold level
}
\keyword{internal}
