% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptune.R
\name{sptune_maxent}
\alias{sptune_maxent}
\title{sptune_maxent}
\usage{
sptune_maxent(x = NULL, p = NULL, data = NULL, nfold = NULL,
  partition_fun = NULL, error_measure = NULL, absence = FALSE,
  beta_multiplier = NULL, feature_classes = NULL, ...)
}
\arguments{
\item{x}{Predictors. Raster* object or SpatialGridDataFrame, containing
grids with predictor variables. These will be used to extract values from
for the point locations. x can also be a data.frame, in which case each
column should be a predictor variable and each row a presence or background
record}

\item{p}{Occurrence data. This can be a data.frame, matrix, SpatialPoints
object, or a vector. If p is a data.frame or matrix it represents a set of
point locations; and it must have two columns with the first being the
x-coordinate (longitude) and the second the y-coordinate (latitude).
Coordinates can also be specified with a SpatialPoints* object}

\item{data}{dataframe.}

\item{nfold}{number of folds for cross-validation.}

\item{partition_fun}{character. Data partitioning method
(e.g. \link{partition_kmeans})}

\item{error_measure}{which error measure to use for optimization.
Default to 'RMSE' for numeric responses, 'AUROC' for binary classification
and 'error' for multiclass classiciation.}

\item{beta_multiplier}{optional user-defined vector of 'beta_multiplier'
hyperparameter to tune over. See details.}

\item{feature_classes}{optional user-defined vector of 'feature_classes'
hyperparameter to tune over. See details.}

\item{...}{additional options passed to \code{partition_fun}.}
}
\description{
Tuning of Maxent ($\beta$ & feature classes) using spatial
cross-validation
}
\details{
This function tunes a Maxent model from the \link{dismo} package
using (spatial) cross-validation.

\code{error_measure} can be specified by the user, selecting one of the returned
error measures of \link{sptune_rf}. However, note that for
regression type responses always the minimum value of the passed error measure
is chosen and for classification cases the highest.

\code{sptune_maxent} is parallelized and runs on all possible cores.
}
\examples{
\dontrun{
library(sperrorest)
data(maxent_pred)
data(maxent_response)
data(basque)

sptune_maxent(x = maxent_pred, p = maxent_response, data = basque,
              absence = TRUE, feature_classes = "QHPT")

out <- sptune_maxent(x = maxent_pred, p = maxent_response, data = basque,
                     absence = TRUE)
}
}
\references{
Brenning, A., Long, S., & Fieguth, P. (2012).
Detecting rock glacier flow structures using Gabor filters and IKONOS
imagery. Remote Sensing of Environment, 125, 227-237.
doi:10.1016/j.rse.2012.07.005

Pena, M. A., & Brenning, A. (2015). Assessing fruit-tree crop classification
from Landsat-8 time series for the Maipo Valley, Chile. Remote Sensing of
Environment, 171, 234-244. doi:10.1016/j.rse.2015.10.029
}
\seealso{
\link{plot_hyper_maxent}
}
\author{
Patrick Schratz, Alexander Brenning
}
