% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptune.R
\name{sptune_rf}
\alias{sptune_rf}
\title{sptune_rf}
\usage{
sptune_rf(formula = NULL, data = NULL, step_factor = 2, nfold = NULL,
  partition_fun = NULL, rf_fun = "rfsrc", error_measure = NULL,
  mtrys = NULL, ntrees = NULL, importance = "none", ...)
}
\arguments{
\item{formula}{formula.}

\item{data}{dataframe.}

\item{step_factor}{option to speed up tuning using less 'ntrees' options.
Default to \code{step_factor = 2}. Increase to reduce number of 'ntrees' for
tuning.}

\item{nfold}{number of folds for cross-validation.}

\item{partition_fun}{character. Data partitioning method
(e.g. \link{partition_kmeans})}

\item{rf_fun}{character. Which R Random Forest package to use. See details.}

\item{error_measure}{which error measure to use for optimization.
Default to 'RMSE' for numeric responses, 'AUROC' for binary classification
and 'error' for multiclass classiciation.}

\item{mtrys}{optional user-defined vector of 'mtry' hyperparameter to
tune over. See details.}

\item{ntrees}{optional user-defined vector of 'ntrees' hyperparameter to
tune over. See details.}

\item{importance}{whether to calculate variable importance on the best
model selected by tuning. See \code{?rfsrc} for more details.}

\item{...}{additional options passed to \code{partition_fun}.}
}
\description{
Tuning of Random Forest (mtry & ntrees) using spatial
cross-validation
}
\details{
This function tunes a Random Forest model either from \link{randomForest},
or \link{randomForestSRC} package using (spatial) cross-validation.

\code{error_measure} can be specified by the user, selecting one of the returned
error measures of \link{sptune_rf}. However, note that for
regression type responses always the minimum value of the passed error measure
is chosen and for classification cases the highest.

The default behaviour of \link{sptune_rf} tunes over all possible 'mtry' values
(which are of \code{length(predictors)}) and a selection of 'ntrees' ranging
between 10 and 2500. Use \code{accelerate} to reduce the number of 'ntrees'.
Specify a custom vector if you want to modify the number of \code{mtry} used
for testing. This is useful if the model contains > 20
predictors but runtime depends on your cpu power / number of cores.

\code{sptune_rf} is parallelized and runs on all possible cores.
}
\examples{
\dontrun{
##------------------------------------------------------------
## binary classification
##------------------------------------------------------------
data(ecuador) # Muenchow et al. (2012), see ?ecuador
fo <- slides ~ dem + slope + hcurv + vcurv + log.carea + cslope

out <- sptune_rf(fo, ecuador, step_factor = 16, nfold = 5,
partition_fun = "partition_kmeans", rf_fun = "randomForest")

##------------------------------------------------------------
## multiclass classification
##------------------------------------------------------------
fo <- croptype ~ b82 + b83 + b84 + b85 + b86 + b87 + ndvi01 +
      ndvi02 + ndvi03 + ndvi04
data(maipo)
out <- sptune_rf(fo, maipo, step_factor = 32, nfold = 5,
                 coords = c("utmx", "utmy"),
                 partition_fun = "partition_kmeans",
                 rf_fun = "randomForest")

##------------------------------------------------------------
## regression
##------------------------------------------------------------

data(ecuador) # Muenchow et al. (2012), see ?ecuador
fo <- dem ~ slides + slope + hcurv + vcurv + log.carea + cslope

out <- sptune_rf(fo, ecuador, step_factor = 16, nfold = 5,
partition_fun = "partition_kmeans", rf_fun = "randomForest")
}
}
\references{
Brenning, A., Long, S., & Fieguth, P. (2012).
Detecting rock glacier flow structures using Gabor filters and IKONOS
imagery. Remote Sensing of Environment, 125, 227-237.
doi:10.1016/j.rse.2012.07.005

Pena, M. A., & Brenning, A. (2015). Assessing fruit-tree crop classification
from Landsat-8 time series for the Maipo Valley, Chile. Remote Sensing of
Environment, 171, 234-244. doi:10.1016/j.rse.2015.10.029
}
\seealso{
\link{plot_hyper_rf}
}
\author{
Patrick Schratz, Alexander Brenning
}
