% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptune.R
\name{sptune_svm}
\alias{sptune_svm}
\title{sptune_svm}
\usage{
sptune_svm(formula = NULL, data = NULL, tuning_parameters = list(),
  train = NULL, test = NULL, average_folds = FALSE, accelerate = 1,
  nfold = NULL, partition_fun = NULL, kernel = NULL, type = NULL,
  error_measure = NULL, svm_fun = "svm", tune = FALSE, ...)
}
\arguments{
\item{formula}{formula.}

\item{data}{dataframe}

\item{accelerate}{option to speed up tuning using less 'cost' and
'gamma' options. Default to \code{accelerate = 1}.
Increase to reduce number of tuning combinations.}

\item{nfold}{number of folds for cross-validation.}

\item{partition_fun}{character. Data partitioning method
(e.g. \link{partition_kmeans})}

\item{kernel}{the kernel to use.}

\item{type}{classification type to use. See \link{svm} or \link{ksvm} for details.}

\item{error_measure}{which error measure to use for optimization.
Default to 'RMSE' for numeric responses, 'AUROC' for binary classification
and 'error' for multiclass classiciation.}

\item{svm_fun}{character. Which R svm package to use. See details.}

\item{...}{additional options passed to \code{partition_fun}.}

\item{tuning_parameter}{named list of tuning parameter containing values
to tune over.}
}
\description{
Tuning of a Support Vector Machine (cost & gamma) using spatial
cross-validation
}
\details{
This function tunes a Support Vector Machine either from \code{e1071}
or \code{kernlab} package using (spatial) cross-validation.

Tuning is performed over the following ranges (\href{https://stats.stackexchange.com/questions/43943/which-search-range-for-determining-svm-optimal-c-and-gamma-parameters}{reference}):
\tabular{rr}{
cost: \tab 2^{-5}, 2^{-3}, ..., 2^{15}\cr
gamma: \tab 2^{-15}, 2^{-13}, ..., 2^{3}
}

\code{error_measure} can be specified by the user, selecting one of the returned
error measures of \link{sptune_svm}. However, note that for
regression type responses always the minimum value of the passed error measure
is chosen and for classification cases the highest.

If \code{parameter1} and \code{parameter2} are unspecified, tuning will be performed
on hyperparameters 'cost' and 'gamma'.

\code{sptune_svm} is parallelized and runs on all possible cores.
}
\examples{
\dontrun{
##------------------------------------------------------------
## binary classification
##------------------------------------------------------------
data(ecuador) # Muenchow et al. (2012), see ?ecuador
fo <- slides ~ dem + slope + hcurv + vcurv + log.carea + cslope

out <- sptune_svm(fo, ecuador, accelerate = 8, nfold = 5,
                  partition_fun = "partition_kmeans", svm_fun = "svm",
                  tuning_parameters = list(gamma = seq(1,5),
                  coef0 = seq(1,4)),
                  kernel = "sigmoid", type = "C-classification")

##------------------------------------------------------------
## multiclass classification
##------------------------------------------------------------
fo <- croptype ~ b82 + b83 + b84 + b85 + b86 + b87 + ndvi01 +
      ndvi02 + ndvi03 + ndvi04
data(maipo)
out <- sptune_svm(fo, maipo, accelerate = 8, nfold = 5,
                  coords = c("utmx", "utmy"),
                  partition_fun = "partition_kmeans",
                  svm_fun = "ksvm", type = "C-svc", kernel = "rbfdot")

##------------------------------------------------------------
## regression
##------------------------------------------------------------

data(ecuador) # Muenchow et al. (2012), see ?ecuador
fo <- dem ~ slides + slope + hcurv + vcurv + log.carea + cslope

out <- sptune_svm(fo, ecuador, accelerate = 8, nfold = 5,
                  partition_fun = "partition_kmeans", svm_fun = "svm",
                  kernel = "radial", type = "eps-regression")
}
}
\references{
Brenning, A., Long, S., & Fieguth, P. (2012).
Detecting rock glacier flow structures using Gabor filters and IKONOS
imagery. Remote Sensing of Environment, 125, 227-237.
doi:10.1016/j.rse.2012.07.005

Pena, M. A., & Brenning, A. (2015). Assessing fruit-tree crop classification
from Landsat-8 time series for the Maipo Valley, Chile. Remote Sensing of
Environment, 171, 234-244. doi:10.1016/j.rse.2015.10.029
}
\seealso{
\link{plot_hyper_svm}
}
\author{
Patrick Schratz, Alexander Brenning
}
