% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary.sperrorestreperror}
\alias{summary.sperrorestreperror}
\alias{summary.sperrorest}
\alias{print.sperrorestimportance}
\alias{print.sperroresterror}
\alias{print.sperrorestreperror}
\alias{print.sperrorest}
\alias{print.sperrorestbenchmarks}
\alias{print.sperrorestpackageversion}
\title{Summary and print methods for sperrorest results}
\usage{
\method{summary}{sperrorestreperror}(object, level = 0, na.rm = TRUE, ...)

\method{summary}{sperrorest}(object, ...)

\method{print}{sperrorestimportance}(x, ...)

\method{print}{sperroresterror}(x, ...)

\method{print}{sperrorestreperror}(x, ...)

\method{print}{sperrorest}(x, ...)

\method{print}{sperrorestbenchmarks}(x, ...)

\method{print}{sperrorestpackageversion}(x, ...)
}
\arguments{
\item{object}{a \link{sperrorest} object}

\item{level}{Level at which errors are summarized:
0: overall; 1: repetition; 2: fold}

\item{na.rm}{Remove \code{NA} values? See \link{mean} etc.}

\item{...}{additional arguments for \link{summary.sperroresterror}
or \link{summary.sperrorestimportance}}

\item{x}{Depending on method, a \link{sperrorest},
\code{sperroresterror} or \code{sperrorestimportance} object}
}
\description{
Summary methods provide varying level of detail while print methods
provide full details.
}
\seealso{
\link{sperrorest},
\link{summary.sperroresterror},
\link{summary.sperrorestimportance}
}
