% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sperrorest.R
\name{summary.sperrorestreperror}
\alias{summary.sperrorestreperror}
\alias{summary.sperrorest}
\alias{print.sperrorestimportance}
\alias{print.sperroresterror}
\alias{print.sperrorestreperror}
\alias{print.sperrorest}
\title{Summary and print methods for sperrorest results}
\usage{
\method{summary}{sperrorestreperror}(object, level = 0, na.rm = TRUE, ...)

\method{summary}{sperrorest}(object, ...)

\method{print}{sperrorestimportance}(x, ...)

\method{print}{sperroresterror}(x, ...)

\method{print}{sperrorestreperror}(x, ...)

\method{print}{sperrorest}(x, ...)
}
\arguments{
\item{object}{a \code{\link[=sperrorest]{sperrorest()}} object}

\item{level}{Level at which errors are summarized:
0: overall; 1: repetition; 2: fold}

\item{na.rm}{Remove \code{NA} values? See \code{\link[=mean]{mean()}} etc.}

\item{...}{additional arguments for \code{\link[=summary.sperroresterror]{summary.sperroresterror()}}
or \code{\link[=summary.sperrorestimportance]{summary.sperrorestimportance()}}}

\item{x}{Depending on method, a \code{\link[=sperrorest]{sperrorest()}},
\code{\link[=sperroresterror]{sperroresterror()}} or \code{\link[=sperrorestimportance]{sperrorestimportance()}} object}
}
\description{
Summary methods provide varying level of detail while print methods
provide full details.
}
\seealso{
\code{\link[=sperrorest]{sperrorest()}}, \code{\link[=sperroresterror]{sperroresterror()}},
\code{\link[=sperrorestimportance]{sperrorestimportance()}},
\code{\link[=summary.sperroresterror]{summary.sperroresterror()}},
\code{\link[=summary.sperrorestimportance]{summary.sperrorestimportance()}}
}
