% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.R
\name{summary.sperrorestbenchmarks}
\alias{summary.sperrorestbenchmarks}
\title{Summarize benchmark information obtained by \code{sperrorest}}
\usage{
\method{summary}{sperrorestbenchmarks}(object, ...)
}
\arguments{
\item{object}{\code{sperrorestbenchmarks} object returned class by
\code{\link[=sperrorest]{sperrorest()}}}

\item{...}{additional arguments (currently ignored)}
}
\value{
List of length seven
}
\description{
\code{summary.sperrorestbenchmarks} shows information on runtime performance,
used cores and system information
}
